calc audioOffset by the difference of duration between audio and video recordings

+++ mainwindow.h
@@ -136,7 +136,7 @@ private:
     void fetchVideo();
 
     QString millisecondsToSecondsString(qint64 milliseconds);
-    int getMediaDuration(const QString &filePath);
+    double getMediaDuration(const QString &filePath);
     void addProgressSong(QGraphicsScene *scene, qint64 duration);
     void updateProgress(const QString& output, QProgressBar* progressBar, int totalDuration);
